+++ app/src/components/DatasetDialog/AddDatasetDialogValidator.test.ts
@@ -1,4 +1,4 @@
-describe('Test validation of AddDatasetDialog', (): void => {
+describe('Test validation of DatasetDialog', (): void => {
   test('Should validate correctly answered dialog', (): void => {
     expect(true).toBeTruthy();
   });
